requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
alembic==1.9.1
python-dotenv==0.21.0
SQLAlchemy==1.4.46
pydantic==1.10.4
